Improve metrics chart
